requirements.txt
pydantic ~= 2.7.4
python-jose[cryptography] ~= 3.3.0
fastapi~=0.111.0
httpx~=0.27.0
tortoise-orm[asyncpg]~=0.21.4
pydantic-settings~=2.3.4
passlib[bcrypt]~=1.7.4
aiofiles~=24.1.0
